requirements.txt
telethon==1.34.0
aiogram==3.4.1
python-dotenv==1.0.0
asyncio
